# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma MOOC API
  version: 0.2.0
  description: >
    API REST de la plataforma web de cursos masivos abiertos en línea.
    Prefijo de versión: /api/v1. Errores uniformes ({"error":{"code","message","details"}}),
    sesiones por cookie httpOnly o cabecera Authorization: Bearer <token>.
    Este documento cubre los endpoints implementados en la iteración actual
    (identidad, administración, autoría de cursos, catálogo e inscripciones).
    Pendientes: quizzes, progreso e insignias (ver README).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /resources/{resourceId}/captions:
    get:
      tags: [Multimedia]
      summary: Pistas de subtítulos de un recurso
      description: >
        Autoriza por la misma vía que el contenido: si el recurso no se le
        puede entregar a quien pregunta, sus subtítulos tampoco. La URL de cada
        pista es la del WebVTT, firmada o servida por el CDN, y es lo que el
        reproductor pone en `<track src>`.
      operationId: listCaptions
      parameters:
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Pistas disponibles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        language: { type: string }
                        label: { type: string }
                        kind: { type: string, enum: [subtitles, captions] }
                        url: { type: string, format: uri }
  /resources/{resourceId}/transcript/{idioma}:
    get:
      tags: [Multimedia]
      summary: Transcripción de un recurso reproducible
      description: >
        La transcripción se deriva de los subtítulos, que es donde ya está el
        texto: pedir las dos cosas por separado garantizaría que acaben
        diciendo cosas distintas. Es lo que hace el material legible para quien
        no puede o no quiere reproducirlo, y lo que permite buscar dentro de
        una clase grabada.
      operationId: getTranscript
      parameters:
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: idioma, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Texto corrido de la pista.
          content:
            application/json:
              schema:
                type: object
                properties:
                  resource_title: { type: string }
                  language: { type: string }
                  transcript: { type: string }
        "404": { $ref: "#/components/responses/Problema" }
  /courses/versions/{versionId}/resources/{resourceId}/captions/{idioma}:
    put:
      tags: [Multimedia]
      summary: Sube o reemplaza la pista de un idioma
      description: >
        El archivo va en el cuerpo JSON y no como carga directa al almacén, al
        contrario que el vídeo, porque el servidor necesita leerlo para derivar
        la transcripción. Reemplaza en lugar de acumular: subir otra vez el
        mismo idioma es corregir la pista, no añadir una segunda que el
        reproductor no sabría cuál elegir.

        Solo se acepta WebVTT. Un SRT es casi idéntico pero el navegador lo
        rechaza, así que aceptarlo aquí solo trasladaría el fallo a la pantalla
        del estudiante.
      operationId: saveCaptions
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: idioma, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vtt]
              properties:
                vtt: { type: string }
                label: { type: string }
                kind: { type: string, enum: [subtitles, captions] }
      responses:
        "200": { description: Pista guardada, con el número de fragmentos y la duración. }
        "422": { $ref: "#/components/responses/Problema" }
    delete:
      tags: [Multimedia]
      summary: Retira la pista de un idioma
      operationId: deleteCaptions
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: idioma, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Pista retirada. }
...
components:
...
  responses:
...
    Problema:
      description: >
        Error uniforme. El campo `error.code` identifica el motivo de forma
        estable; `error.details` enumera los motivos cuando hay varios, como en
        la validación de publicación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mooc_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
...